[feat] 펫, 품종 엔티티 생성 및 관계 설정 (#20)
* [feat] 펫, 품종 엔티티 생성 및 관계 설정

* [chore] 파일 이동

* [feat] 펫 생성 구현 및 기타 수정

* [feat] 품종 이름 enum 추가

* [chore] 파일 삭제

* [feat] 펫 삭제, 내 펫 목록 조회 API 구현 및 테스트

* [feat] 내 펫 목록 조회 테스트

+++ src/main/kotlin/com/example/fatpetserver/common/auth/UserDetails.kt
@@ -8,7 +8,7 @@ data class UserDetails(
     private val _authorities: List<SimpleGrantedAuthority>,
 ) : UserDetails {
 
-    val id = _id
+    val id = _id.toLong()
 
     override fun getAuthorities() = _authorities
